requirements.txt
apache-airflow
apache-airflow-providers-amazon
apache-airflow-providers-cncf-kubernetes
boto3
pytest
pytest-cov
pandas
moto